semester check

+++ modules/db/src/main/scala/gem/dao/SemesterDao.scala
@@ -23,7 +23,7 @@ object SemesterDao {
     def canonicalize(s: Semester): Update0 =
       sql"""
         INSERT INTO semester (semester_id, year, half)
-        VALUES (${s.toString}, ${s.getYear}, ${s.getHalf})
+        VALUES (${s.toString}, ${s.getYear.toShort}, ${s.getHalf})
         ON CONFLICT DO NOTHING
       """.update
